LeNet-5/requirements.txt
numpy>=1.17.0
torch>=1.4.0
torchvision>=0.4.0
visdom>=0.1.6
Pillow~=11.0.0
onnx~=1.18.0

onnxruntime~=1.20.1
matplotlib~=3.7.2